agent-wallet/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openclaw-agent-wallet"
version = "0.1.0"
description = "Plugin-friendly wallet backend for OpenClaw agents"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pynacl>=1.5.0",
    "python-dotenv>=1.0.0",
    "solana>=0.36.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent_wallet"]

[tool.ruff]
target-version = "py311"
line-length = 100
